refactor: 놓인 돌의 타입 변경
