+++ README.md
@@ -7,6 +7,7 @@
 - 돌 
   - 돌은 위치와 타입을 가진다.
 - 오목판
+  - 놓인 돌을 가진다.
   - 돌을 추가한다
 - 턴
   - 돌을 둘 수 있는지 확인한다.
