+++ src/main/kotlin/omok/controller/OmokController.kt
@@ -3,10 +3,10 @@ package omok.controller
 import omok.domain.Game
 import omok.domain.model.Board
 import omok.domain.model.position.Column
+import omok.domain.model.position.OmokStone
 import omok.domain.model.position.Position
 import omok.domain.model.position.Row
 import omok.domain.model.state.BlackStoneTurn
-import omok.domain.model.stone.OmokStone
 import omok.domain.model.stone.StoneType
 import omok.view.InputView
 import omok.view.OutputView
