+++ src/main/kotlin/omok/domain/Game.kt
@@ -1,32 +1,30 @@
 package omok.domain
 
 import omok.domain.model.Board
+import omok.domain.model.position.OmokStone
 import omok.domain.model.position.Position
 import omok.domain.model.state.Finish
 import omok.domain.model.state.OmokState
 import omok.domain.model.state.Running
-import omok.domain.model.stone.OmokStone
 import omok.domain.model.stone.StoneType
 
 class Game(private val state: OmokState) {
     fun play(
         onBeforePlace: (Board, StoneType, OmokStone?) -> Unit,
         onPlace: () -> Position,
-    ): OmokStone {
-        return play(state, onBeforePlace, onPlace).let {
-            it.lastStoneOrNull() ?: error("게임이 종료되지 않았습니다.")
-        }
+    ): StoneType {
+        return play(state, onBeforePlace, onPlace).stoneType
     }
 
     private tailrec fun play(
         state: OmokState,
         onBeforePlace: (Board, StoneType, OmokStone?) -> Unit,
         onPlace: () -> Position,
-    ): Board {
+    ): OmokState {
         onBeforePlace(state.board, state.stoneType, state.board.lastStoneOrNull())
         return when (val newState = state.placeStone(onPlace)) {
             is Running -> play(newState, onBeforePlace, onPlace)
-            is Finish -> newState.board
+            is Finish -> newState
         }
     }
 }
