+++ src/main/kotlin/omok/domain/model/position/OmokStone.kt
@@ -1,6 +1,6 @@
-package omok.domain.model.stone
+package omok.domain.model.position
 
-import omok.domain.model.position.Position
+import omok.domain.model.stone.StoneType
 
 data class OmokStone(
     val position: Position,
